requirements.txt
edt==2.3.0
gputools==0.2.13
h5py==3.8.0
imageio==2.29.0
matplotlib==3.1.3
numba==0.56.0
numexpr==2.8.4
numpy==1.21.6
pandas==1.1.3
Pillow==9.5.0
PyYAML==6.0
scipy==1.7.3
setuptools==67.7.2
six==1.16.0
scikit-image==0.19.3
tifffile==2020.6.3
winregistry==1.1.1
torch==1.13.1
torchvision==0.14.1
tqdm==4.62.3